use C type in prop
git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile@725 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ Services/Private/Azalea/AZClient+GNUstep.m
@@ -19,7 +19,7 @@
   unsigned long *data;
   unsigned int num;
   if (!prop_get_array32(window, prop_atoms.gnustep_wm_attr,
-		  prop_atoms.gnustep_wm_attr, (unsigned int **)&data, &num))
+		  prop_atoms.gnustep_wm_attr, (unsigned long **)&data, &num))
     return;
 
   if (num != 9) {
